Fix #15 - allow to run multiple requests in the same test
It turns out (at least in Slim 2.6) that the request (and the response)
are a singleton in the Slim app.
Thus even when mocking a new \Slim\Environment like the WebTestclient does,
you'll always get the previous request that points to the previous
environment (and thus the previous endpoint).

The idea is to create fresh request and response object on every client
call, and also override the matched route cache that is built into
the router.

Signed-off-by: Brice Figureau <brice@daysofwonder.com>